python_driver/pyproject.toml
[project]
name = "esp-now-transponder-python"
version = "0.1.0"
description = "Python interface for using a USB dongle as an ESP-NOW transponder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipdb>=0.13.13",
    "pygame>=2.6.1",
    "pyserial>=3.5",
]
